Implement M3 and M4 campaigns and align narrative artwork
Add Radar Warning and Airlift mission flows, cinematic camera tours,
bilingual comics, tutorials, ARIA guidance, field guides and persistence.
Align earlier comics and commander/ARIA portraits with the M3 art style.
Repair resource HUD values, Persian text, result visibility/layout and
transport presentation. Include production plans, architecture records
and text Editor QA evidence. Design-folder review images remain local.

Validation: M4 full Editor acceptance passed all 14 regression suites
and the live rescue, settlement, replay, retry and deadline-failure flow.
The final vehicle-panel repair passed a fresh real rescue; eight result
and 56 comic presentation captures passed. Existing architecture guard
failures and the unreproduced APC fragmentation report remain documented.

+++ Assets/Game/Scripts/UI/Screens/MatchHudAssistantUiSystemHelper.M02PlacementGuidance.cs
@@ -35,8 +35,8 @@ namespace Game.UI.Runtime
             public readonly byte DisplayedTutorialStep;
             public readonly UiTutorialNarrationPhase PendingTutorialPhase;
             public readonly UiTutorialNarrationPhase DisplayedTutorialPhase;
-            public readonly ushort NarratedTutorialCues;
-            public readonly ushort AutoShownTutorialCues;
+            public readonly uint NarratedTutorialCues;
+            public readonly uint AutoShownTutorialCues;
             public readonly float TutorialShowAt;
             public readonly bool TutorialCinematicSuspended;
             public readonly bool FinalTutorialSuppressed;
